Add gtm file
