+++ web/app/themes/justicejobs/webpack.mix.js
@@ -10,6 +10,7 @@ mix.js([
         'src/js/map.js',
         'src/js/main.js'
     ], dist + 'js/main.min.js')
+    .js('src/js/jj-gtm.js', dist + 'js/jj-gtm.min.js')
     .sass('src/scss/style.scss', dist + 'css/main.min.css')
     .copy('src/video/*', dist + 'video')
     .copy('src/fonts/*', dist + 'fonts')
